pyproject.toml
[tool.poetry]
name = "wikiviews"
version = "0.1.0"
description = ""
authors = ["sarangj <sarangjoshi22@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
Flask = "2.2.2"
python-dateutil = "2.8.2"
requests = "2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
pytest-mock = "3.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
